@@ -18,15 +18,15 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "aws-amplify": "^5.3.26",
+    "aws-amplify": "^6.6.0",
     "amazon-connect-chatjs": "^2.3.0",
-    "@aws-sdk/client-cognito-identity": "3.470.0",
     "@aws-sdk/client-connect": "3.470.0",
     "@aws-sdk/client-lex-runtime-service": "3.470.0",
     "@aws-sdk/client-lex-runtime-v2": "3.470.0",
     "@aws-sdk/client-polly": "3.470.0",
     "@aws-sdk/client-s3": "3.470.0",
-    "@aws-sdk/credential-providers": "3.470.0",
+    "@smithy/signature-v4": "^5.0.0",
+    "@aws-crypto/sha256-js": "^5.2.0",
     "browserify-zlib": "^0.2.0",
     "buffer": "^6.0.3",
     "core-js": "^3.6.5",
@@ -36,10 +36,10 @@
     "node-polyfill-webpack-plugin": "^2.0.1",
     "roboto-fontface": "^0.10.0",
     "stream-browserify": "^3.0.0",
-    "vue": "^3.3.10",
+    "vue": "^3.5.0",
     "vue-loader": "^17.3.1",
-    "vue-router": "^4.2.5",
-    "vuetify": "^3.4.6",
+    "vue-router": "^4.5.1",
+    "vuetify": "^3.8.0",
     "vuex": "^4.1.0"
   },
   "devDependencies": {
